Code And Such
This is most of the player movement done along with a follow camera. slopes are wonky but fixable if I take time on it. Bombs still need to be implemented allong with an attack and a dash. I also need to insert some placeholder sprites so that we can see which direction the sprite is facing and make sure the dash goes that way.